pyproject.toml
[tool.poetry]
name = "discord-tiny-image"
version = "0.1.0"
description = "Discord bot which converts an input text into an image with tiny text."
authors = ["Lilith Roth <lilith@roth.systems>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = "^2.4.0"
pillow = "<=9.5.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
